requirements.txt
flask
pytest
redis
celery
gevent
flower
requests
Flask-Pydantic
flask-sqlalchemy
faker
pydantic[email]
prometheus-client
mlflow
pytest-cov
pydantic
flask_sqlalchemy
prometheus_client
pydantic[email]
email-validator
